add create route for vote

+++ pkg/dto/voteDTO.go
@@ -6,10 +6,10 @@ type VoteData struct {
 	UUID        uuid.UUID `json:"id"`
 	UserId      uuid.UUID `json:"userId"`
 	CountryId   uuid.UUID `json:"countryId"`
-	Costume     int       `json:"costume"`
-	Song        int       `json:"song"`
-	Performance int       `json:"performance"`
-	Props       int       `json:"props"`
+	Costume     int8      `json:"costume"`
+	Song        int8      `json:"song"`
+	Performance int8      `json:"performance"`
+	Props       int8      `json:"props"`
 }
 
 type Vote struct {
